4.21*: bloglist expansion, step9

+++ part4/Blog-List/controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -9,17 +10,21 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.get('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
-      if (blog) {
-        response.json(blog)
-      } else {
-        response.status(404).end()
-      }
-    })
+  if (blog) {
+    response.json(blog)
+  } else {
+    response.status(404).end()
+  }
+})
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
 
-  const user = await User.find({})
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
   if (body.title === undefined || body.author === undefined) {
     return response.status(400).json({ error: 'bad request' })
@@ -30,21 +35,34 @@ blogsRouter.post('/', async (request, response) => {
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
-    user: user[0].id
+    user: user.id
   })
 
   const savedBlog = await blog.save()
-  console.log(savedBlog.id);
-  console.log(user.blogs);
-  user[0].blogs = user[0].blogs.concat(savedBlog._id)
-  await user[0].save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
 
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const body = request.body
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(400).json({ error: 'no permission' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
